[TIMOB-17876] 64 bit safe code migration
Turn on GCC_WARN_64_TO_32_BIT_CONVERSION flag and fix code for the same

+++ iphone/Classes/SCListener.m
@@ -184,7 +184,7 @@ static void listeningCallback(void *inUserData, AudioQueueRef inAQ, AudioQueueBu
 	return self;
 }
 
-- (unsigned)retainCount {
+- (NSUInteger)retainCount {
 	return UINT_MAX;
 }
